# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FichoChat API",
        "description": "API REST versionnée de FichoChat. Chaque endpoint documente des exemples JSON complets de réponses (succès et erreurs).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/devices": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Lister les appareils",
                "operationId": "devicesIndex",
                "responses": {
                    "200": {
                        "description": "Liste des appareils",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DevicesIndex200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Enregistrer un appareil",
                "description": "Le client envoie uniquement la clé publique. Le premier appareil est auto-approuvé ; les suivants restent pending.",
                "operationId": "devicesRegister",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "platform",
                                    "public_key"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "MacBook Pro"
                                    },
                                    "platform": {
                                        "type": "string",
                                        "example": "web",
                                        "enum": [
                                            "web",
                                            "ios",
                                            "android",
                                            "desktop"
                                        ]
                                    },
                                    "public_key": {
                                        "type": "string",
                                        "example": "-----BEGIN PUBLIC KEY-----\\nMFkwEwYH...\\n-----END PUBLIC KEY-----"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "name": "MacBook Pro",
                                "platform": "web",
                                "public_key": "-----BEGIN PUBLIC KEY-----\\nMFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAExamplePublicKeyDataHere==\\n-----END PUBLIC KEY-----"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appareil enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DevicesRegister201"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/devices/{id}": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Détail d’un appareil",
                "operationId": "devicesShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appareil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DevicesShow200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound404"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/devices/{id}/approve": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Approuver un appareil",
                "description": "Doit être appelé depuis un appareil déjà approuvé (actor_device_id).",
                "operationId": "devicesApprove",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "actor_device_id"
                                ],
                                "properties": {
                                    "actor_device_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "actor_device_id": "019fc7e1-aaaa-bbbb-cccc-ddddeeeeffff"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appareil approuvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DevicesApprove200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound404"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation / règles métier",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/devices/{id}/revoke": {
            "post": {
                "tags": [
                    "Devices"
                ],
                "summary": "Révoquer un appareil",
                "description": "Révoque l’appareil et invalide ses refresh tokens.",
                "operationId": "devicesRevoke",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "actor_device_id"
                                ],
                                "properties": {
                                    "actor_device_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "actor_device_id": "019fc7e1-aaaa-bbbb-cccc-ddddeeeeffff"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appareil révoqué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DevicesRevoke200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorNotFound404"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation / règles métier",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation422"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/devices/pending": {
            "get": {
                "tags": [
                    "Devices"
                ],
                "summary": "Appareils en attente",
                "operationId": "devicesPending",
                "responses": {
                    "200": {
                        "description": "Appareils pending",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DevicesPending200"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorUnauthorized401"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorServer500"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorUnauthorized401": {
                "description": "Non authentifié",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object",
                "example": {
                    "success": false,
                    "message": "Unauthenticated."
                }
            },
...
            "ErrorServer500": {
                "description": "Erreur serveur",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Internal server error."
                    }
                },
                "type": "object",
                "example": {
                    "success": false,
                    "message": "Internal server error."
                }
            },
...
            "DevicesIndex200": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Operation completed successfully.",
                    "data": {
                        "devices": [
                            {
                                "id": "019fc7e1-aaaa-bbbb-cccc-ddddeeeeffff",
                                "name": "MacBook Pro",
                                "platform": "web",
                                "fingerprint": "a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef123456",
                                "public_key": "-----BEGIN PUBLIC KEY-----\\nMFkwEwYH...\\n-----END PUBLIC KEY-----",
                                "status": "approved",
                                "approved_at": "2026-08-03T18:00:00.000000Z",
                                "approved_by_device_id": null,
                                "revoked_at": null,
                                "last_seen_at": "2026-08-03T18:05:00.000000Z",
                                "created_at": "2026-08-03T18:00:00.000000Z",
                                "updated_at": "2026-08-03T18:05:00.000000Z"
                            }
                        ]
                    }
                }
            },
            "DevicesRegister201": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Device registered successfully.",
                    "data": {
                        "device": {
                            "id": "019fc7e1-aaaa-bbbb-cccc-ddddeeeeffff",
                            "name": "MacBook Pro",
                            "platform": "web",
                            "fingerprint": "a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef123456",
                            "public_key": "-----BEGIN PUBLIC KEY-----\\nMFkwEwYH...\\n-----END PUBLIC KEY-----",
                            "status": "approved",
                            "approved_at": "2026-08-03T18:00:00.000000Z",
                            "approved_by_device_id": null,
                            "revoked_at": null,
                            "last_seen_at": "2026-08-03T18:00:00.000000Z",
                            "created_at": "2026-08-03T18:00:00.000000Z",
                            "updated_at": "2026-08-03T18:00:00.000000Z"
                        }
                    }
                }
            },
            "DevicesShow200": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Operation completed successfully.",
                    "data": {
                        "device": {
                            "id": "019fc7e1-aaaa-bbbb-cccc-ddddeeeeffff",
                            "name": "iPhone 15",
                            "platform": "ios",
                            "fingerprint": "bbccddeeff00112233445566778899aabbccddeeff00112233445566778899aa",
                            "public_key": "-----BEGIN PUBLIC KEY-----\\nMFkwEwYH...\\n-----END PUBLIC KEY-----",
                            "status": "pending",
                            "approved_at": null,
                            "approved_by_device_id": null,
                            "revoked_at": null,
                            "last_seen_at": "2026-08-03T18:10:00.000000Z",
                            "created_at": "2026-08-03T18:10:00.000000Z",
                            "updated_at": "2026-08-03T18:10:00.000000Z"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Access Token obtenu via /auth/login ou /auth/register. Header: Authorization: Bearer {access_token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}